+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Create parameter data classes (`Param` and `Struct`)
 - Create database class `ParamDB` to store parameters in a SQLite file
-- Add ability to retrieve the commit history as `CommitEntry` objects.
+- Add ability to retrieve the commit history as `CommitEntry` objects
+- Create [documentation website](https://painterqubits.github.io/paramdb)
 
 [unreleased]: https://github.com/PainterQubits/paramdb